Update the heroku command to be able to allow atomic deploy (#46)
The previous version only allowed to deploy one dyno type at a time. For
projects with multiple dyno types, it is problematic.

This change is backward incompatible and replace the old CLI API with a new
one allowing to pass multiple dyno-type:artifact mapping. For example:

```
taskboot deploy-heroku --heroku-app bugbug \
    "web:public/bugbug/bugbug-http-service.tar" \
    "worker:public/bugbug/bugbug-http-service-bg-worker.tar"
```

All the artifact are retrieved, pushed and then only one call is made to the Heroku
API to deploy them atomically.